notebook experimenting with different ways of getting tau vs z profiles from simulation.
In this notebook, i experiment with an approach that allows me to interpolate between different tau vs z  profiles based on the value of the intensity. This interpolation works reasonably well however, it's not significanly better than Kclustering. Even with this 'smoother' interpolation method, I'm left with certain discontinuities that I'm not sure how to deal with....