experiment to see how sensitive the spherical tensors are to perpendicular velocities
I do a simple experiment where I take a section of a MURaM simulation and calculate the resulting spherical tensors. THen I zero out the x and y component of velcoties in the atmosphere and recompute the spherical tensors.

I find that the resulting spherical tensors are not significantly different. Hence, not being able to invert for perpendiculaar components of velocity won't be too much of a problem here.